GERONIMO-2965 GERONIMO-3008 GERONIMO-3010 Tomcat annotation processing with lots of other improvements.  Uses LifecycleProvider interface proposed in GERONIMO-3010
git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@522036 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/geronimo-jetty6/src/main/java/org/apache/geronimo/jetty6/JettyFilterHolder.java
@@ -51,11 +51,11 @@ public class JettyFilterHolder implements GBeanLifecycle {
     }
 
     public void doStart() throws Exception {
-            filterHolder.start();
+        filterHolder.start();
     }
 
     public void doStop() throws Exception {
-                filterHolder.stop();
+        filterHolder.stop();
     }
 
     public void doFail() {
@@ -68,6 +68,7 @@ public class JettyFilterHolder implements GBeanLifecycle {
 
     private static class InternalFilterHolder extends FilterHolder {
         private final JettyServletRegistration servletRegistration;
+        private boolean destroyed;
 
         public InternalFilterHolder(JettyServletRegistration servletRegistration) {
             this.servletRegistration = servletRegistration;
@@ -79,12 +80,15 @@ public class JettyFilterHolder implements GBeanLifecycle {
         }
 
         public void destroyInstance(Object o) throws Exception {
-            super.destroyInstance(o);
-            servletRegistration.destroyInstance(o);
+            if (!destroyed) {
+                super.destroyInstance(o);
+                servletRegistration.destroyInstance(o);
+                destroyed = true;
+            }
         }
 
     }
-    
+
     public static final GBeanInfo GBEAN_INFO;
 
     static {
@@ -95,7 +99,7 @@ public class JettyFilterHolder implements GBeanLifecycle {
 
         infoBuilder.addReference("JettyServletRegistration", JettyServletRegistration.class, NameFactory.WEB_MODULE);
 
-        infoBuilder.setConstructor(new String[] {"filterName", "filterClass", "initParams", "JettyServletRegistration"});
+        infoBuilder.setConstructor(new String[]{"filterName", "filterClass", "initParams", "JettyServletRegistration"});
 
         GBEAN_INFO = infoBuilder.getBeanInfo();
